+++ src/wavwriter.hpp
@@ -13,6 +13,6 @@ enum SampleFmt {
 
 const char* toString(SampleFmt format);
 
-void writewav(float *data, SampleFmt format, int num_channels, int samples, int sample_rate, const char* filename);
+void writewav(uint8_t *data, SampleFmt format, int num_channels, int samples, int sample_rate, const char* filename);
 
 #endif
